+++ tests/test_jupyter_hook.py
@@ -3,6 +3,7 @@ import os
 import nbrmd
 from .utils import list_all_notebooks, remove_outputs, \
     remove_outputs_and_header
+from nbrmd.cm import update_alternative_formats
 
 
 @pytest.mark.parametrize('nb_file', list_all_notebooks('.ipynb'))
@@ -11,7 +12,10 @@ def test_rmd_is_ok(nb_file, tmpdir):
     tmp_ipynb = str(tmpdir.join('notebook.ipynb'))
     tmp_rmd = str(tmpdir.join('notebook.Rmd'))
 
-    nbrmd.update_rmd(model=dict(type='notebook', content=nb), path=tmp_ipynb)
+    nb.metadata['nbrmd_formats'] = ['.Rmd']
+    update_alternative_formats(
+        model=dict(type='notebook', content=nb),
+        path=tmp_ipynb)
 
     nb2 = nbrmd.readf(tmp_rmd)
 
@@ -24,7 +28,9 @@ def test_ipynb_is_ok(nb_file, tmpdir):
     tmp_ipynb = str(tmpdir.join('notebook.ipynb'))
     tmp_rmd = str(tmpdir.join('notebook.Rmd'))
 
-    nbrmd.update_ipynb(model=dict(type='notebook', content=nb), path=tmp_rmd)
+    update_alternative_formats(
+        model=dict(type='notebook', content=nb),
+        path=tmp_rmd)
 
     nb2 = nbrmd.readf(tmp_ipynb)
 
@@ -39,8 +45,9 @@ def test_all_files_created(nb_file, tmpdir):
     tmp_rmd = str(tmpdir.join('notebook.Rmd'))
     nb.metadata['nbrmd_formats'] = ['.Rmd', '.ipynb', '.md']
 
-    nbrmd.update_selected_formats(
-        model=dict(type='notebook', content=nb), path=tmp_ipynb)
+    update_alternative_formats(
+        model=dict(type='notebook', content=nb),
+        path=tmp_ipynb)
 
     nb2 = nbrmd.readf(tmp_md)
     assert remove_outputs_and_header(nb) == remove_outputs_and_header(nb2)
@@ -54,7 +61,7 @@ def test_no_files_created_on_no_format(tmpdir):
     tmp_md = str(tmpdir.join('notebook.md'))
     tmp_rmd = str(tmpdir.join('notebook.Rmd'))
 
-    nbrmd.update_selected_formats(
+    update_alternative_formats(
         model=dict(type='notebook',
                    content=dict(nbformat=4, metadata=dict())),
         path=tmp_ipynb)
@@ -67,7 +74,7 @@ def test_raise_on_wrong_format(tmpdir):
     tmp_ipynb = str(tmpdir.join('notebook.ipynb'))
 
     with pytest.raises(TypeError):
-        nbrmd.update_selected_formats(
+        update_alternative_formats(
             model=dict(type='notebook',
                        content=dict(nbformat=4,
                                     metadata=dict(nbrmd_formats=['.doc']))),
@@ -78,7 +85,8 @@ def test_no_rmd_on_not_notebook(tmpdir):
     tmp_ipynb = str(tmpdir.join('notebook.ipynb'))
     tmp_rmd = str(tmpdir.join('notebook.Rmd'))
 
-    nbrmd.update_rmd(model=dict(type='not notebook'), path=tmp_ipynb)
+    update_alternative_formats(model=dict(type='not notebook'),
+                               path=tmp_ipynb)
     assert not os.path.isfile(tmp_rmd)
 
 
@@ -86,7 +94,8 @@ def test_no_rmd_on_not_v4(tmpdir):
     tmp_ipynb = str(tmpdir.join('notebook.ipynb'))
     tmp_rmd = str(tmpdir.join('notebook.Rmd'))
 
-    nbrmd.update_rmd(
-        model=dict(type='notebook', content=dict(nbformat=3)), path=tmp_ipynb)
+    update_alternative_formats(
+        model=dict(type='notebook', content=dict(nbformat=3)),
+        path=tmp_rmd)
 
-    assert not os.path.isfile(tmp_rmd)
+    assert not os.path.isfile(tmp_ipynb)